Implementacion de menu con productos y clientes

+++ demo/src/main/java/com/example/demo/controller/ClientePerfilController.java
@@ -38,7 +38,7 @@ public class ClientePerfilController {
 
     // Ver perfil: Cuando queremos ver el perfil de un cliente con su id
     @GetMapping("/ver/{id}")
-    public String verPerfil(@PathVariable Integer id, Model model) { // PathVariable integer id nos permite que si la
+    public String verPerfil(@PathVariable Long id, Model model) { // PathVariable Long id nos permite que si la
                                                                      // URL es perfil/ver/5 se combierte a id=5
         Cliente cliente = clienteService.obtenerPorId(id); // Le pedimos al servicio a clienteService que busque al
                                                            // cliente con ese ID en el repositorio
@@ -53,7 +53,7 @@ public class ClientePerfilController {
 
     // Edigtar perfil: Cuando el cliente quiera cambiar algunos de sus datos
     @GetMapping("/editar/{id}")
-    public String mostrarEditar(@PathVariable Integer id, Model model) { // Toma el numero que viene con la URL
+    public String mostrarEditar(@PathVariable Long id, Model model) { // Toma el numero que viene con la URL
         Cliente cliente = clienteService.obtenerPorId(id); // Se busca a traves del service el cliente por su id
         if (cliente == null) { // Si el cliente no existe o el id es invalido lo redirigimos al registtro
             return "redirect:/perfil/registro";
@@ -73,7 +73,7 @@ public class ClientePerfilController {
     // Desactivar cuenta: Le damos la opcion al cliente para que desactive su
     // cuenta, mas no la elimine
     @GetMapping("/desactivar/{id}")
-    public String desactivar(@PathVariable Integer id) { // Toma el numero que viene con la URL
+    public String desactivar(@PathVariable Long id) { // Toma el numero que viene con la URL
         clienteService.desactivar(id); // Le pedimos al service que busque al cliente y lo desactive
         return "redirect:/perfil/registro"; // Lo mandamos a la pagina de registro
     }
@@ -81,14 +81,14 @@ public class ClientePerfilController {
     // Reactivar la cuenta: Despues de desactivar la cuenta el cliente tiene la
     // opcion de volver a activarla si desea
     @GetMapping("/reactivar/{id}")
-    public String reactivar(@PathVariable Integer id) {// Toma el numero que viene con la URL
+    public String reactivar(@PathVariable Long id) {// Toma el numero que viene con la URL
         clienteService.reactivar(id); // Le pedimos al service que busque al cliente y lo vuelva a activar
         return "redirect:/perfil/ver/" + id; // Lo redirigimos a su perfil para que lo pueda ver
     }
 
     // Elimina la cuenta(Borra la cuenta): Cuando el cliente quiere borrar su cuenta
     @GetMapping("/eliminar/{id}")
-    public String eliminar(@PathVariable Integer id) { // Toma el numero que viene con la URL
+    public String eliminar(@PathVariable Long id) { // Toma el numero que viene con la URL
         clienteService.eliminar(id); // Le pedimos al servicio que elimine la cuenta del cliente por su id
         return "redirect:/perfil/registro"; // Redirigimos al cliente a resgistro
     }
