+++ demo/src/main/java/com/example/demo/service/ClienteService.java
@@ -8,16 +8,16 @@ public interface ClienteService {
 
     // === MÉTODOS DE LECTURA ===
     Collection<Cliente> obtenerTodos();
-    Cliente obtenerPorId(Integer id);
-    boolean existe(Integer id);
+    Cliente obtenerPorId(Long id);
+    boolean existe(Long id);
 
     // === CRUD BÁSICO ===
     Cliente guardar(Cliente cliente);
-    void eliminar(Integer id);
+    void eliminar(Long id);
 
     // === LÓGICA DE NEGOCIO (antes estaba en el controller) ===
     Cliente registrar(Cliente cliente);      // guarda un cliente nuevo y lo activa
     Cliente actualizar(Cliente cliente);     // conserva la contraseña anterior si no se envía una nueva
-    void desactivar(Integer id);
-    void reactivar(Integer id);
+    void desactivar(Long id);
+    void reactivar(Long id);
 }
