+++ demo/src/main/java/com/example/demo/service/ProductoService.java
@@ -7,10 +7,10 @@ import com.example.demo.entities.Producto;
 public interface ProductoService {
     // === MÉTODOS DE LECTURA (YA EXISTENTES) ===
     Collection<Producto> obtenerTodos();
-    Producto obtenerPorId(Integer id);
+    Producto obtenerPorId(Long id);
 
     // === NUEVOS MÉTODOS PARA CRUD ===
     Producto guardar(Producto producto);
-    void eliminar(Integer id);
-    boolean existe(Integer id);
+    void eliminar(Long id);
+    boolean existe(Long id);
 }
